// generated/src/app/api-models/ipa-afflt-addr-contact.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class IpaAffltAddrContact {
  seqIpaAffltAddrCtat: number;
  seqIpaAffltAddress: number;
  contactName: string;
  contactTitle: string;
  phoneNumber: string;
  extension: string;
  faxNumber: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}